% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_select.R
\name{index_laslogs}
\alias{index_laslogs}
\title{Build a FAIR index for a folder of LAS files}
\usage{
index_laslogs(dir)
}
\arguments{
\item{dir}{Folder containing .las files}
}
\value{
A list with wells_index, curves_index, files_index
}
\description{
Build a FAIR index for a folder of LAS files
}
