% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_header.R
\name{parse_header_block}
\alias{parse_header_block}
\title{Parse LAS header-style lines (MNEM.UNIT VALUE : DESC)}
\usage{
parse_header_block(lines)
}
\description{
Parse LAS header-style lines (MNEM.UNIT VALUE : DESC)
}
\keyword{internal}
