% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_laslog.R
\name{read_laslog}
\alias{read_laslog}
\title{Read a LAS well log file (Log ASCII Standard) into a structured object}
\usage{
read_laslog(file, output = c("long", "wide"))
}
\arguments{
\item{file}{Path to a .las file}

\item{output}{Output format for LOG: "long" (tidy) or "wide"}
}
\value{
An S3 object of class "laslog" with VERSION/WELL/CURVE/PARAMETER/OTHER/LOG
}
\description{
Read a LAS well log file (Log ASCII Standard) into a structured object
}
