% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_select.R
\name{read_laslog_header}
\alias{read_laslog_header}
\title{Read LAS header only (no ~A data)}
\usage{
read_laslog_header(file)
}
\arguments{
\item{file}{Path to a .las file}
}
\value{
S3 object of class "laslog_header" with VERSION/WELL/CURVE/PARAMETER/OTHER plus provenance
}
\description{
Read LAS header only (no ~A data)
}
