% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_write.R
\name{pull_laslogs}
\alias{pull_laslogs}
\title{Pull log data for selected wells (optionally selected curves)}
\usage{
pull_laslogs(index, apis, curves = NULL, output = c("long", "wide"))
}
\arguments{
\item{index}{Output of index_laslogs()}

\item{apis}{Character vector of API values to load}

\item{curves}{Optional curve mnemonics to keep (e.g., c("GR","RHOB","NPHI"))}

\item{output}{"long" (tidy) or "wide" (ML-ready)}
}
\value{
A tibble combining all selected wells
}
\description{
Pull log data for selected wells (optionally selected curves)
}
