% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_select.R
\name{select_laslogs}
\alias{select_laslogs}
\title{Select wells from an index by metadata and curve availability}
\usage{
select_laslogs(index, county = NULL, curves_any = NULL, curves_all = NULL)
}
\arguments{
\item{index}{Output of index_laslogs()}

\item{county}{Character vector of counties to keep (optional)}

\item{curves_any}{Keep wells that have at least one of these curves (optional)}

\item{curves_all}{Keep wells that have all of these curves (optional)}
}
\value{
Character vector of API values
}
\description{
Select wells from an index by metadata and curve availability
}
