% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_write.R
\name{write_laslogs}
\alias{write_laslogs}
\title{Write LAS logs to CSV and/or Parquet}
\usage{
write_laslogs(data, out_dir, prefix = "laslogs", csv = TRUE, parquet = TRUE)
}
\arguments{
\item{data}{Tibble returned by pull_laslogs()}

\item{out_dir}{Output directory}

\item{prefix}{File prefix (no extension)}

\item{csv}{Write CSV file?}

\item{parquet}{Write Parquet file? (requires arrow)}
}
\value{
Invisibly returns output paths
}
\description{
Write LAS logs to CSV and/or Parquet
}
